Move tests of dof permuter to their own file, and fix their meshes
